removing datatypes
