+++ client-src/elements/external-reviewers.ts
@@ -2,7 +2,11 @@
 
 /** Represents an external organization that reviews Chromium changes. Currently the W3C TAG,
  * Firefox, and Safari. */
-import {LabelInfo} from './datatypes.js';
+type LabelInfo = {
+  description: string;
+  variant: 'primary' | 'success' | 'neutral' | 'warning' | 'danger';
+};
+
 export class ExternalReviewer {
   icon: string;
   _labels: Record<string, LabelInfo>;
